model_server/requirements.txt
rai-core-flask==0.5.0
raiutils>=0.4.0
